Transaction.Cost to type=decimal

+++ OnlineMarketPlace.Models/TransactionDetail.cs
@@ -17,7 +17,7 @@ namespace OnlineMarketPlace.Models
             PayPal
         }
         public int Id { get; set; }
-        public double Cost { get; set; }
+        public decimal Cost { get; set; }
         [Display(Name = "Created")]
         public DateTimeOffset CreatedUtc { get; set; }
         [Display(Name = "Modified")]
